+++ lib/repositories/task_repository_interface.dart
@@ -1,5 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter_todolist_app/Models/Task.dart';
+
+import 'package:flutter_todolist_app/models/task.dart';
 
 abstract class TaskRepositoryInterface {
   Future create(Task task);
